Unified login/logout methods
