+++ src/huggingface_hub/_login.py
@@ -1,0 +1,229 @@
+# Copyright 2020 The HuggingFace Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Contains methods to login to the Hub."""
+import subprocess
+from getpass import getpass
+from typing import List, Optional
+
+from .commands._cli_utils import ANSI
+from .hf_api import HfApi
+from .utils import HfFolder, is_google_colab, is_notebook, logging, run_subprocess
+
+
+logger = logging.get_logger(__name__)
+
+
+def login(token: Optional[str] = None) -> None:
+    """Login the machine to access the Hub.
+
+    If `token` is provided, it is persisted in cache and set as a git credential. Once
+    done, the machine is logged in and the access token will be available across all
+    `huggingface_hub` components. If `token` is not provided, it will be prompted to the
+    user either with a widget (in a notebook) or via the terminal.
+
+    To login from outside of a script, one can also use `huggingface-cli login` which is
+    a cli command that wraps [`login`].
+
+    Args:
+        token (`str`, *optional*):
+            User access token to generate from https://huggingface.co/settings/token.
+
+    Raises:
+        [`ValueError`](https://docs.python.org/3/library/exceptions.html#ValueError)
+            If an organization token is passed. Only personal account tokens are valid
+            to login.
+        [`ValueError`](https://docs.python.org/3/library/exceptions.html#ValueError)
+            If token is invalid.
+        [`ImportError`](https://docs.python.org/3/library/exceptions.html#ImportError)
+            If running in a notebook but `ipywidgets` is not installed.
+    """
+    if token is not None:
+        _login(token)
+    elif is_notebook():
+        notebook_login()
+    else:
+        interpreter_login()
+
+
+def logout() -> None:
+    """Logout the machine from the Hub.
+
+    Token is deleted from the machine and removed from git credential.
+    """
+    token = HfFolder.get_token()
+    if token is None:
+        print("Not logged in!")
+        return
+    HfFolder.delete_token()
+    HfApi.unset_access_token()
+    print("Successfully logged out.")
+
+
+###
+# Interpreter-based login (text)
+###
+
+
+def interpreter_login():
+    """
+    Displays a prompt to login to the HF website and store the token.
+    """
+    print(  # docstyle-ignore
+        """
+    _|    _|  _|    _|    _|_|_|    _|_|_|  _|_|_|  _|      _|    _|_|_|      _|_|_|_|    _|_|      _|_|_|  _|_|_|_|
+    _|    _|  _|    _|  _|        _|          _|    _|_|    _|  _|            _|        _|    _|  _|        _|
+    _|_|_|_|  _|    _|  _|  _|_|  _|  _|_|    _|    _|  _|  _|  _|  _|_|      _|_|_|    _|_|_|_|  _|        _|_|_|
+    _|    _|  _|    _|  _|    _|  _|    _|    _|    _|    _|_|  _|    _|      _|        _|    _|  _|        _|
+    _|    _|    _|_|      _|_|_|    _|_|_|  _|_|_|  _|      _|    _|_|_|      _|        _|    _|    _|_|_|  _|_|_|_|
+
+    To login, `huggingface_hub` now requires a token generated from https://huggingface.co/settings/tokens .
+    """
+    )
+    _login(token=getpass("Token: "))
+
+
+###
+# Notebook-based login (widget)
+###
+
+NOTEBOOK_LOGIN_PASSWORD_HTML = """<center> <img
+src=https://huggingface.co/front/assets/huggingface_logo-noborder.svg
+alt='Hugging Face'> <br> Immediately click login after typing your password or
+it might be stored in plain text in this notebook file. </center>"""
+
+
+NOTEBOOK_LOGIN_TOKEN_HTML_START = """<center> <img
+src=https://huggingface.co/front/assets/huggingface_logo-noborder.svg
+alt='Hugging Face'> <br> Copy a token from <a
+href="https://huggingface.co/settings/tokens" target="_blank">your Hugging Face
+tokens page</a> and paste it below. <br> Immediately click login after copying
+your token or it might be stored in plain text in this notebook file. </center>"""
+
+
+NOTEBOOK_LOGIN_TOKEN_HTML_END = """
+<b>Pro Tip:</b> If you don't already have one, you can create a dedicated
+'notebooks' token with 'write' access, that you can then easily reuse for all
+notebooks. </center>"""
+
+
+def notebook_login():
+    """
+    Displays a widget to login to the HF website and store the token.
+    """
+    try:
+        import ipywidgets.widgets as widgets
+        from IPython.display import clear_output, display
+    except ImportError:
+        raise ImportError(
+            "The `notebook_login` function can only be used in a notebook (Jupyter or"
+            " Colab) and you need the `ipywidgets` module: `pip install ipywidgets`."
+        )
+
+    box_layout = widgets.Layout(
+        display="flex", flex_flow="column", align_items="center", width="50%"
+    )
+
+    token_widget = widgets.Password(description="Token:")
+    token_finish_button = widgets.Button(description="Login")
+
+    login_token_widget = widgets.VBox(
+        [
+            widgets.HTML(NOTEBOOK_LOGIN_TOKEN_HTML_START),
+            token_widget,
+            token_finish_button,
+            widgets.HTML(NOTEBOOK_LOGIN_TOKEN_HTML_END),
+        ],
+        layout=box_layout,
+    )
+    display(login_token_widget)
+
+    # On click events
+    def login_token_event(t):
+        token = token_widget.value
+        # Erase token and clear value to make sure it's not saved in the notebook.
+        token_widget.value = ""
+        clear_output()
+        _login(token)
+
+    token_finish_button.on_click(login_token_event)
+
+
+###
+# Login private helpers
+###
+
+
+def _login(token: str) -> None:
+    hf_api = HfApi()
+    if token.startswith("api_org"):
+        raise ValueError("You must use your personal account token.")
+    if not hf_api._is_valid_token(token=token):
+        raise ValueError("Invalid token passed!")
+    hf_api.set_access_token(token)
+    HfFolder.save_token(token)
+    print("Login successful")
+    print("Your token has been saved to", HfFolder.path_token)
+
+    # Only in Google Colab to avoid the warning message
+    # See https://github.com/huggingface/huggingface_hub/issues/1043#issuecomment-1247010710
+    if is_google_colab():
+        _set_store_as_git_credential_helper_globally()
+
+    helpers = _currently_setup_credential_helpers()
+
+    if "store" not in helpers:
+        print(
+            ANSI.red(
+                "Authenticated through git-credential store but this isn't the helper"
+                " defined on your machine.\nYou might have to re-authenticate when"
+                " pushing to the Hugging Face Hub. Run the following command in your"
+                " terminal in case you want to set this credential helper as the"
+                " default\n\ngit config --global credential.helper store"
+            )
+        )
+
+
+def _set_store_as_git_credential_helper_globally() -> None:
+    """Set globally the credential.helper to `store`.
+
+    To be used only in Google Colab as we assume the user doesn't care about the git
+    credential config. It is the only particular case where we don't want to display the
+    warning message in `notebook_login()`.
+
+    Related:
+    - https://github.com/huggingface/huggingface_hub/issues/1043
+    - https://github.com/huggingface/huggingface_hub/issues/1051
+    - https://git-scm.com/docs/git-credential-store
+    """
+    try:
+        run_subprocess("git config --global credential.helper store")
+    except subprocess.CalledProcessError as exc:
+        raise EnvironmentError(exc.stderr)
+
+
+def _currently_setup_credential_helpers(directory=None) -> List[str]:
+    try:
+        output = run_subprocess(
+            "git config --list".split(),
+            directory,
+        ).stdout.split("\n")
+
+        current_credential_helpers = []
+        for line in output:
+            if "credential.helper" in line:
+                current_credential_helpers.append(line.split("=")[-1])
+    except subprocess.CalledProcessError as exc:
+        raise EnvironmentError(exc.stderr)
+
+    return current_credential_helpers
